refactor: massively reorganise headers and includes, replace a great number of primative types

+++ include/FLCore/X86Math/DllMain.cpp
@@ -1,6 +1,6 @@
 #include <Windows.h>
 
-BOOL WINAPI DllMain(HINSTANCE hinstDLL, unsigned long fdwReason, LPVOID lpvReserved)
+BOOL WINAPI DllMain(HINSTANCE hinstDLL, unsigned fdwReason, LPVOID lpvReserved)
 {
     if (fdwReason == DLL_PROCESS_ATTACH)
     {
